requirements.txt
anki==25.2.5
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
decorator==5.2.1
distro==1.9.0
dotenv==0.9.9
google-api-core==2.25.1
google-api-python-client==2.172.0
google-api-python-client-stubs==1.29.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-texttospeech==2.27.0
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.10.0
Markdown==3.8
openai==1.82.0
orjson==3.10.18
proto-plus==1.26.1
protobuf==6.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.7
tqdm==4.67.1
types-httplib2==0.22.0.20250516
typing-inspection==0.4.1
typing_extensions==4.13.2
uritemplate==4.2.0
urllib3==2.4.0
websockets==15.0.1
